backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
python-dotenv==1.0.1
pyjwt==2.9.0
psycopg2-binary==2.9.10
bcrypt==4.0.1
passlib[bcrypt]>=1.7.4
email-validator==2.2.0
cohere==5.9.3